# Conflicts:
#	.idea/.gitignore
#	.idea/vcs.xml
#	course-project-feedme.iml
#	src/main/java/Entity/Order.java
#	src/main/java/Entity/Restaurant.java
#	src/test/java/UserTest.java

+++ src/main/java/Entity/Product.java
@@ -8,7 +8,7 @@ public class Product {
     private int stock;
 
     // Initialize Product
-    public Product(String name, String id, float price, int stock){
+    public Product(String name, String id, double price, int stock){
         this.name = name;
         this.id = id;
         this.price = price;
